fix circular deps in react-table
