+++ packages/react-table/src/components/Table/index.ts
@@ -1,4 +1,3 @@
-export * from './Table';
 export * from './ActionsColumn';
 export * from './Body';
 export * from './BodyCell';
@@ -12,5 +11,7 @@ export * from './HeaderCell';
 export * from './RowWrapper';
 export * from './SelectColumn';
 export * from './SortColumn';
+export * from './Table';
+export * from './TableContext';
 export * from './TableText';
 export * from './utils';
