+++ packages/react-table/src/components/Table/utils/decorators/selectable.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { css } from '@patternfly/react-styles';
 import styles from '@patternfly/react-styles/css/components/Table/table';
-import { IExtra, IFormatterValueType, ITransform, RowSelectVariant } from '../../Table';
+import { IExtra, IFormatterValueType, ITransform, RowSelectVariant } from '../../TableContext';
 import { SelectColumn } from '../../SelectColumn';
 import checkStyles from '@patternfly/react-styles/css/components/Check/check';
