+++ packages/react-table/src/components/Table/utils/decorators/textCenter.ts
@@ -1,2 +1,2 @@
-import { ITransform } from '../../Table';
+import { ITransform } from '../../TableContext';
 export const textCenter: ITransform = () => ({ textCenter: true });
