+++ packages/react-table/src/components/Table/utils/formatters.tsx
@@ -1,4 +1,4 @@
-import { IFormatterValueType, IFormatter } from '../Table';
+import { IFormatterValueType, IFormatter } from '../TableContext';
 
 const defaultTitle: IFormatter = (data: IFormatterValueType) =>
   data && data.hasOwnProperty('title') ? data.title : data;
