+++ packages/react-table/src/components/Table/utils/transformers.tsx
@@ -11,7 +11,7 @@ export { headerCol } from './decorators/headerCol';
 export { classNames, Visibility } from './decorators/classNames';
 export { info } from './decorators/info';
 
-import { IFormatterValueType, IExtra, ITransform } from '../Table';
+import { IFormatterValueType, IExtra, ITransform } from '../TableContext';
 
 const emptyTD: ITransform = () => ({
   scope: '',
