+++ packages/react-table/src/components/Table/utils/utils.tsx
@@ -1,4 +1,4 @@
-import { IRow, IRowCell, IValidatorDef, RowEditType, RowErrors } from '../Table';
+import { IRow, IRowCell, IValidatorDef, RowEditType, RowErrors } from '../TableContext';
 
 const hasCompoundParentsExpanded = (parentId: number, compoundParent: number, rows: IRow[]) => {
   // max rows.length parents
