+++ packages/react-table/src/test-helpers/data-helpers.tsx
@@ -1,4 +1,4 @@
-import { IRow } from '../components/Table/Table';
+import { IRow } from '../components';
 
 export const buildExpandableRows = (relationships: any = {}, openIndexes: number[] = [], rowCount = 10) => {
   const rows: IRow[] = [];
